+++ Tests/SentryTests/Integrations/MetricKit/SentryMetricKitIntegrationTests.swift
@@ -39,27 +39,22 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
     }
 
     func testOptionEnabled_MetricKitManagerInitialized() {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             let sut = SentryMetricKitIntegration()
             
             givenInstalledWithEnabled(sut)
             
             XCTAssertNotNil(Dynamic(sut).metricKitManager as SentryMXManager?)
-        }
     }
     
     func testOptionDisabled_MetricKitManagerNotInitialized() {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             let sut = SentryMetricKitIntegration()
             
             sut.install(with: Options())
             
             XCTAssertNil(Dynamic(sut).metricKitManager as SentryMXManager?)
-        }
     }
     
     func testUninstall_MetricKitManagerSetToNil() {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             let sut = SentryMetricKitIntegration()
             
             let options = Options()
@@ -68,11 +63,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             sut.uninstall()
             
             XCTAssertNil(Dynamic(sut).metricKitManager as SentryMXManager?)
-        }
     }
     
     func testMXCrashPayloadReceived() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -82,11 +75,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             mxDelegate.didReceiveCrashDiagnostic(MXCrashDiagnostic(), callStackTree: callStackTreePerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
             
             try assertPerThread(exceptionType: "MXCrashDiagnostic", exceptionValue: "MachException Type:(null) Code:(null) Signal:(null)", exceptionMechanism: "MXCrashDiagnostic", handled: false)
-        }
     }
     
     func testAttachDiagnosticAsAttachment() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -101,11 +92,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
                 
                 XCTAssertEqual(diagnosticAttachment?.data, diagnostic.jsonRepresentation())
             }
-        }
     }
     
     func testDontAttachDiagnosticAsAttachment() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -119,11 +108,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
                 
                 XCTAssertNil(diagnosticAttachment)
             }
-        }
     }
     
     func testSetInAppIncludes_AppliesInAppToStackTrace() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -141,11 +128,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
                 
                 XCTAssertEqual(2, inAppFramesCount)
             }
-        }
     }
     
     func testCPUExceptionDiagnostic_PerThread() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -155,11 +140,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             mxDelegate.didReceiveCpuExceptionDiagnostic(TestMXCPUExceptionDiagnostic(), callStackTree: callStackTreePerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
             
             assertNothingCaptured()
-        }
     }
     
     func testCPUExceptionDiagnostic_NotPerThread() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -169,11 +152,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             mxDelegate.didReceiveCpuExceptionDiagnostic(TestMXCPUExceptionDiagnostic(), callStackTree: callStackTreeNotPerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
             
             try assertNotPerThread(exceptionType: "MXCPUException", exceptionValue: "MXCPUException totalCPUTime:2.2 ms totalSampledTime:5.5 ms", exceptionMechanism: "mx_cpu_exception")
-        }
     }
     
     func testCPUExceptionDiagnostic_OnlyOneFrame() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -201,16 +182,13 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
                 
                 XCTAssertEqual(1, sentryFrames.count)
                 let frame = sentryFrames.first
-                XCTAssertNil(frame?.function)
                 XCTAssertEqual("0x000000021f1a0001", frame?.imageAddress)
                 XCTAssertEqual("libsystem_pthread.dylib", frame?.package)
                 XCTAssertFalse(frame?.inApp?.boolValue ?? true)
             }
-        }
     }
     
     func testDiskWriteExceptionDiagnostic() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -220,11 +198,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             mxDelegate.didReceiveDiskWriteExceptionDiagnostic(TestMXDiskWriteExceptionDiagnostic(), callStackTree: callStackTreeNotPerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
             
             try assertNotPerThread(exceptionType: "MXDiskWriteException", exceptionValue: "MXDiskWriteException totalWritesCaused:5.5 Mib", exceptionMechanism: "mx_disk_write_exception")
-        }
     }
     
     func testHangDiagnostic() throws {
-        if #available(iOS 15, macOS 12, macCatalyst 15, *) {
             givenSDKWithHubWithScope()
             
             let sut = SentryMetricKitIntegration()
@@ -234,10 +210,8 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             mxDelegate.didReceiveHangDiagnostic(TestMXHangDiagnostic(), callStackTree: callStackTreeNotPerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
             
             try assertNotPerThread(exceptionType: "MXHangDiagnostic", exceptionValue: "MXHangDiagnostic hangDuration:6.6 sec", exceptionMechanism: "mx_hang_diagnostic")
-        }
     }
     
-    @available(iOS 15, macOS 12, macCatalyst 15, *)
     private func givenInstalledWithEnabled(_ integration: SentryMetricKitIntegration, optionsBlock: (Options) -> Void = { _ in }) {
         let options = Options()
         options.enableMetricKit = true
@@ -383,7 +357,6 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
   
 }
 
-@available(iOS 15, macOS 12, macCatalyst 15, *)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 class TestMXCallStackTree: MXCallStackTree {
@@ -398,7 +371,6 @@ class TestMXCallStackTree: MXCallStackTree {
     }
 }
 
-@available(iOS 15, macOS 12, macCatalyst 15, *)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 class TestMXCrashDiagnostic: MXCrashDiagnostic {
@@ -413,7 +385,6 @@ class TestMXCrashDiagnostic: MXCrashDiagnostic {
     }
 }
 
-@available(iOS 15, macOS 12, macCatalyst 15, *)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 class TestMXCPUExceptionDiagnostic: MXCPUExceptionDiagnostic {
@@ -436,7 +407,6 @@ class TestMXCPUExceptionDiagnostic: MXCPUExceptionDiagnostic {
     }
 }
 
-@available(iOS 15, macOS 12, macCatalyst 15, *)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 class TestMXDiskWriteExceptionDiagnostic: MXDiskWriteExceptionDiagnostic {
@@ -455,7 +425,6 @@ class TestMXDiskWriteExceptionDiagnostic: MXDiskWriteExceptionDiagnostic {
     }
 }
 
-@available(iOS 15, macOS 12, macCatalyst 15, *)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 class TestMXHangDiagnostic: MXHangDiagnostic {